pyproject.toml
[tool.poetry]
name = "gurkbot"
version = "0.1.0"
description = "Our community bot, used for running the server."
authors = ["The Gurkult Community <noreply@gurkult.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/gurkult/gurkbot"
repository = "https://github.com/gurkult/gurkbot"
packages = [
    { include = "bot" },
]

[tool.poetry.dependencies]
python = "^3.9.0"
disnake = "^2.3"
loguru = "^0.5.3"
PyYAML = "^5.4.1"
Pillow = "^8.2.0"
fuzzywuzzy = "^0.18.0"
asyncpg = "^0.23.0"
python-dateutil = "^2.8.2"
humanize = "^3.11.0"
mcstatus = "^6.5.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"
isort = "^5.10.1"
black = "^21.6b0"
pep8-naming = "^0.11.1"
pre-commit = "^2.13.0"
taskipy = "^1.8.1"
python-dotenv = "^0.18.0"

[tool.taskipy.tasks]
bot = { cmd = "python -m bot", help = "Runs Bot"}
lint = { cmd = "pre-commit run --all-files", help = "Lints project" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black --check .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
